+++ packages/core/src/common/logger/index.ts
@@ -1,2 +1,3 @@
 export * from './logger-protocol';
 export * from './logger';
+export * from './abstract-logger';