+++ packages/core/src/common/logger/logger.ts
@@ -1,35 +1,31 @@
-import * as log from 'loglevel';
-import { Component, Value } from '../annotation';
-import { Logger, LOGGER_CONFIG } from './logger-protocol';
+import { Component } from '../annotation';
+import { Scope } from '../container';
+import { AbstractLogger } from './abstract-logger';
+import { Logger } from './logger-protocol';
 
-@Component(Logger)
-export class LoggerImpl implements Logger {
-    protected instance: Logger;
+@Component({ id: Logger, scope: Scope.Transient })
+export class LoggerImpl extends AbstractLogger {
 
-    constructor(
-        @Value(LOGGER_CONFIG) protected readonly config: any
-    ) {
-        if (config?.level) {
-            log.setLevel(config.level);
-        } else {
-            log.setLevel('error');
+    error(message: any, context?: string): void {
+        this.call(console.error.bind(console), message, context);
+    }
+
+    info(message: any, context?: string): void {
+        if (['info', 'debug'].includes(this.level)) {
+            this.call(console.info.bind(console), message, context);
         }
     }
 
-    error(message: any, context = ''): void {
-        log.error(message, context);
+    warn(message: any, context?: string): void {
+        if (['info', 'debug', 'warn'].includes(this.level)) {
+            this.call(console.warn.bind(console), message, context);
+        }
     }
 
-    info(message: any, context = ''): void {
-        log.info(message, context);
-    }
-
-    warn(message: any, context = ''): void {
-        log.warn(message, context);
-    }
-
-    debug(message: any, context = ''): void {
-        log.debug(message, context);
+    debug(message: any, context?: string): void {
+        if (this.level === 'debug') {
+            this.call(console.debug.bind(console), message, context);
+        }
     }
 
 }