+++ packages/logger/src/common/index.ts
@@ -1,1 +1,2 @@
-export * from './logger';
+export * from './winston-logger';
+export * from './winston-logger-protocol';