+++ packages/logger/src/common/module.ts
@@ -1,4 +1,4 @@
-import '../common';
+import './index';
 import { autoBind } from '@malagu/core';
 
 export default autoBind();