+++ packages/logger/src/common/winston-logger-protocol.ts
@@ -1,0 +1,13 @@
+import { Logger, LoggerOptions } from 'winston';
+
+export const RawWinstonLogger = Symbol('WinstonLogger');
+
+export const WinstonConfig = Symbol('WinstonConfig');
+
+export interface WinstonConfig extends LoggerOptions {
+
+}
+
+export interface RawWinstonLogger extends Logger {
+
+}