+++ packages/logger/src/node/module.ts
@@ -1,4 +1,0 @@
-import '../common';
-import { autoBind } from '@malagu/core';
-
-export default autoBind();