+++ packages/web/src/node/trace/index.ts
@@ -1,3 +1,4 @@
 export * from './trace-middleware';
 export * from './trace-protocol';
 export * from './trace-id-resolver';
+export * from './trace-id-provider';