+++ packages/web/src/node/trace/trace-id-provider.ts
@@ -1,0 +1,10 @@
+import { Component, TraceIdProvider } from '@malagu/core';
+import { Context } from '../context';
+
+@Component(TraceIdProvider)
+export class TraceIdProviderImpl implements TraceIdProvider {
+
+    provide(): string | undefined {
+        return Context.getTraceId();
+    }
+}